+++ app/javascript/src/components/Insurances.vue
@@ -177,14 +177,15 @@
 import { onMounted, ref } from 'vue'
 import VPagination from '@hennge/vue3-pagination'
 import '@hennge/vue3-pagination/dist/vue3-pagination.css'
-import Swal from 'sweetalert2'
 import LoadingAnimation from './LoadingAnimation'
 import { useFormat } from '../composables/useFormat'
 import { useInsurances } from '../composables/useInsurances'
+import { useToast } from '../composables/useToast'
 
 const { formatYen, formatPercent } = useFormat()
 const { insurances, totalPages, getInsurances, deleteInsurance } =
   useInsurances()
+const { toast } = useToast()
 
 const pageParam = () => {
   const url = new URL(location.href)
@@ -206,17 +207,6 @@ const switchPage = (pageNum) => {
   currentPage.value = pageNum
   history.pushState(null, null, newUrl)
   getInsurances(newParams)
-}
-
-const toast = (title) => {
-  Swal.fire({
-    title: title,
-    toast: true,
-    position: 'top-end',
-    showConfirmButton: false,
-    timer: 3000,
-    timerProgressBar: true
-  })
 }
 
 const destroy = async (insuranceId) => {
